+++ pylivecoinwatch/api.py
@@ -4,6 +4,7 @@ import json
 import warnings
 
 from requests.adapters import HTTPAdapter
+
 # from requests.packages.urllib3.util.retry import Retry
 
 # from .utils import func_args_preprocessing
@@ -11,40 +12,41 @@ SANDBOX_API_KEY = "NO_API"
 
 
 class LiveCoinWatchAPI:
-    base_url = 'https://api.livecoinwatch.com'
+    base_url = "https://api.livecoinwatch.com"
 
     def __init__(self, api_key=None):
         self.api_base_url = LiveCoinWatchAPI.base_url
         self.request_timeout = 120
         self.headers = {
-            'content-type': 'application/json',
-            'x-api-key': SANDBOX_API_KEY if api_key == None else api_key
+            "content-type": "application/json",
+            "x-api-key": SANDBOX_API_KEY if api_key == None else api_key,
         }
         self.session = requests.Session()
-        self.session.mount('http://', HTTPAdapter(max_retries=5))
+        self.session.mount("http://", HTTPAdapter(max_retries=5))
         self.session.headers.update(self.headers)
 
     def set_api_key(self, user_api_key):
-        self.headers['x-api-key'] = user_api_key
+        self.headers["x-api-key"] = user_api_key
         self.session.headers.update(self.headers)
 
     def __request(self, url, payload):
         url = "{}/{}".format(self.api_base_url, url)
         try:
-            response = self.session.post(url, data=json.dumps(
-                payload), timeout=self.request_timeout)
+            response = self.session.post(
+                url, data=json.dumps(payload), timeout=self.request_timeout
+            )
         except requests.exceptions.RequestException:
             raise
 
         try:
             # if(response.status_code == 500 and self.__request('status', {}).status_code == 200):
             response.raise_for_status()
-            content = json.loads(response.content.decode('utf-8'))
+            content = json.loads(response.content.decode("utf-8"))
             return content
         except Exception as e:
             # check if json (with error message) is returned
             try:
-                content = json.loads(response.content.decode('utf-8'))
+                content = json.loads(response.content.decode("utf-8"))
                 raise ValueError(content)
             # if no json
             except json.decoder.JSONDecodeError:
@@ -53,39 +55,39 @@ class LiveCoinWatchAPI:
             raise
 
     def status(self):
-        url = 'status'
+        url = "status"
         payload = {}
         return self.__request(url, payload)
 
     def credits(self):
-        url = 'credits'
+        url = "credits"
         payload = {}
         return self.__request(url, payload)
 
     def overview(self, **kwargs):
-        url = 'overview'
+        url = "overview"
         payload = kwargs
         return self.__request(url, payload)
 
     def overview_history(self, start, end, **kwargs):
-        url = 'overview/history'
+        url = "overview/history"
         payload = kwargs
-        payload['start'] = start
-        payload['end'] = end
+        payload["start"] = start
+        payload["end"] = end
         return self.__request(url, payload)
 
     def coins_single(self, code, **kwargs):
-        url = 'coins/single'
+        url = "coins/single"
         payload = kwargs
-        payload['code'] = code
+        payload["code"] = code
         return self.__request(url, payload)
 
     def coins_single_history(self, code, start, end, **kwargs):
-        url = 'coins/single/history'
+        url = "coins/single/history"
         payload = kwargs
-        payload['code'] = code
-        payload['start'] = start
-        payload['end'] = end
+        payload["code"] = code
+        payload["start"] = start
+        payload["end"] = end
         return self.__request(url, payload)
 
     def coins_list(self, **kwargs):
@@ -99,12 +101,30 @@ class LiveCoinWatchAPI:
         return self.__request(url, payload)
 
     def exchanges_single(self, code, **kwargs):
-        url = 'exchanges/single'
+        url = "exchanges/single"
         payload = kwargs
-        payload['code'] = code
+        payload["code"] = code
         return self.__request(url, payload)
 
     def exchanges_list(self, **kwargs):
-        url = 'exchanges/list'
+        url = "exchanges/list"
         payload = kwargs
+        return self.__request(url, payload)
+
+    def coins_contract(self, platform, address, **kwargs):
+        url = "coins/contract"
+        payload = kwargs
+        payload["platform"] = platform
+        payload["address"] = address
+        return self.__request(url, payload)
+
+    def platforms_all(self):
+        url = "platforms/all"
+        payload = {}
+        return self.__request(url, payload)
+
+    def coins_map(self, codes, **kwargs):
+        url = "coins/map"
+        payload = kwargs
+        payload["codes"] = codes
         return self.__request(url, payload)